[PATCH] Upstream patch - 17022023

See merge request flectra-hq/flectra!943

+++ addons/resource/models/resource.py
@@ -12,7 +12,7 @@ from pytz import timezone, utc
 
 from flectra import api, fields, models, _
 from flectra.addons.base.models.res_partner import _tz_get
-from flectra.exceptions import ValidationError
+from flectra.exceptions import ValidationError, UserError
 from flectra.osv import expression
 from flectra.tools.float_utils import float_round
 
@@ -285,6 +285,9 @@ class ResourceCalendar(models.Model):
         self.hours_per_day = self._compute_hours_per_day(attendances)
 
     def switch_calendar_type(self):
+        if self == self.env.company.resource_calendar_id:
+            raise UserError(_('Impossible to switch calendar type for the default company schedule.'))
+
         if not self.two_weeks_calendar:
             self.attendance_ids.unlink()
             self.attendance_ids = [
